functionality to delete individual items from orders

+++ src/main/java/com/qa/ims/controller/OrdersController.java
@@ -60,7 +60,7 @@ public class OrdersController implements CrudController<Orders> {
 			LOGGER.info("Please enter the quanitity you would like of this item");
 			Long quantity = utils.getLong();
 			
-			OrdersItems ordersItems = oi.create(new OrdersItems(orders.getId(), itemId , quantity));
+			OrdersItems ordersItems = oi.create(new OrdersItems(orders.getOrderId(), itemId , quantity));
 			LOGGER.info("Item added to order");
 	
 			LOGGER.info("Would you like to order another item? Y/N?");
@@ -90,9 +90,26 @@ public class OrdersController implements CrudController<Orders> {
 
 	@Override
 	public int delete() {
+		LOGGER.info("Please enter the number of the option you would like"
+				+ "\n [1] delete an entire order from the system" +
+				"\n [2] delete an item from an order");
+		int response = scan.nextInt();
+		if(response == 1) {
 		LOGGER.info("Please enter the id of the order you would like to delete");
 		Long id = utils.getLong();
-		return ordersDAO.delete(id);
+		int orders = ordersDAO.delete(id);
+		return ordersDAO.deletePhaseTwo(id);
+		}
+		else if(response == 2) {
+			LOGGER.info("Please enter the id of the order from which you would like to delete an item");
+			Long id = utils.getLong();
+			LOGGER.info("Please enter the id of the item you wish to delete");
+			Long itemId = utils.getLong();
+			return oi.deleteItem(id, itemId);
+		} else {
+			 LOGGER.info("That is not a valid selection");
+		}
+	 return 0;
 	}
